requirements.txt
msgpack-python
tornado
autobahntestsuite
websocket-client
humanize
psutil
